Refactor Welcome section: reorganize components and add new sections for benefits, use cases, and FAQs. Remove unused ActivityTrack components and update imports accordingly.
