+++ src/components/Welcome/ApplyCasesSection.tsx
