+++ src/components/Welcome/BenefitsSection.tsx
