+++ src/components/Welcome/CoreFeaturesSection.tsx
