+++ src/components/Welcome/FAQSection.tsx
@@ -23,10 +23,6 @@ export default function FAQSection() {
       answer: "Simply take a photo of items in your fridge or pantry with your smartphone camera. Our AI-powered image recognition technology identifies the food items and provides nutritional information instantly."
     },
     {
-      question: "Is NutriPeek suitable for families with food allergies?",
-      answer: "Absolutely! You can set up dietary preferences and restrictions including allergies, and NutriPeek will provide safe recommendations and alerts for potential allergens in food items."
-    },
-    {
       question: "How does the plate builder feature help my child?",
       answer: "The interactive plate builder helps children learn good eating habits by visually showing what a balanced meal looks like. It makes healthy eating fun and educational, allowing kids to participate in creating their own balanced meals."
     },
@@ -37,6 +33,26 @@ export default function FAQSection() {
     {
       question: "How can NutriPeek help with my morning time constraints?",
       answer: "NutriPeek streamlines the lunch preparation process by quickly identifying what you have available, suggesting balanced meal combinations, and providing time-saving recommendations based on your morning schedule."
+    },
+    {
+      question: "How will the match & learn feature help my child to learn more about food and be interested in them?",
+      answer: "The match & learn feature makes food education engaging through interactive games and colorful visuals. Children can discover where foods come from, learn about their nutritional benefits, and complete fun food-matching challenges. This gamified approach nurtures curiosity about healthy eating and helps children develop positive relationships with nutritious foods."
+    },
+    {
+      question: "How are the ingredients being recommended to me?",
+      answer: "Our recommendation system identifies nutritional gaps in your meal plan and searches our database for ingredients rich in those missing nutrients. We analyze your current ingredients, determine which food groups need supplementing, and suggest options that provide the highest nutritional value while complementing your existing food items."
+    },
+    {
+      question: "What is METy?",
+      answer: "METy (Youth Metabolic Equivalent) is a measurement unit for children's energy expenditure during physical activities. Unlike adult METs, METy values are specially adjusted for children's higher metabolic rates. One METy represents a child's basal metabolic rate, and activities are rated as multiples of this baseline. NutriPeek uses age-appropriate METy values to accurately assess your child's energy needs."
+    },
+    {
+      question: "How is the Physical Activity Level (PAL) calculated?",
+      answer: "Your PAL is calculated by tracking your daily activities and their intensity levels. We multiply the time spent on each activity by its corresponding METy value, sum these values, and divide by the total minutes in a day. This gives us an accurate picture of your overall physical activity, which helps determine your energy requirements."
+    },
+    {
+      question: "How is the estimated energy requirement based on PAL calculated?",
+      answer: "Your child's estimated energy requirement is calculated by mapping their age, gender, and physical activity level (PAL) to established nutritional guidelines. We use data from the Australian Dietary Guidelines to determine the appropriate energy needs based on these factors. This provides a personalized daily calorie target that supports healthy growth and development while accounting for their unique activity patterns."
     }
   ];
