+++ src/components/Welcome/FooterSection.tsx
