+++ src/components/Welcome/HeroSection.tsx
