+++ src/components/Welcome/PlateBuilderSection.tsx
