+++ src/components/Welcome/SectionContainer.tsx
