+++ src/components/Welcome/SectionIndicator.tsx
