+++ src/components/Welcome/ToolsIntegrationSection.tsx
